pyproject.toml
[project]
name = "translation-app"
version = "0.1.0"
description = "LangChain Translation App - Webpage EN to ZH-TW translator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit",
    "langchain",
    "langchain-openai",
    "langchain-text-splitters",
    "langchain-core",
    "requests",
    "beautifulsoup4",
    "python-dotenv",
    "lxml",
    "langchain-community>=0.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
